web/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
constantly==15.1.0
cryptography==40.0.0
daphne==4.0.0
Django==4.1.7
djangorestframework==3.14.0
gunicorn==21.2.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
packaging==23.1
priority==1.3.0
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.1.0
pytz==2022.7.1
redis==4.5.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
zope.interface==6.0
